refactor: rename providers to contexts across components
Replaced `providers` with `contexts` across multiple components for consistency with established naming conventions. Updated all related imports, exports, file paths, and references to reflect these changes in tasks, workspaces, inbox, and projects.
