+++ apps/nextjs/src/components/pages/inbox/components/activity/activity.tsx
@@ -17,7 +17,7 @@ import { TasksContext } from '@/components/features/tasks/tasks-provider/tasks-c
 import { Flex } from '@/components/ui/flex';
 import { useInboxActivityPageQuery } from '@/hooks/queries/app';
 import { getInboxDetailId, isInboxDetailURL } from '@/router';
-import { useInboxPageContext } from '../../providers/context';
+import { useInboxPageContext } from '../../contexts/context';
 
 export const Activity = memo(function Activity() {
   return <Component />;
