+++ apps/nextjs/src/components/pages/projects/component.tsx
@@ -22,7 +22,7 @@ import { Files } from './components/files';
 import { Header } from './components/header';
 import { List } from './components/list';
 import { Overview } from './components/overview';
-import { Context, useProjectsPageContext } from './providers/context';
+import { Context, useProjectsPageContext } from './contexts/context';
 
 type Props = {
   loading: boolean;
