+++ apps/nextjs/src/components/pages/projects/components/files/files.tsx
@@ -4,7 +4,7 @@ import { TasksContainer } from '@/components/features/tasks/tasks-container/task
 import { TasksFilesContent } from '@/components/features/tasks/tasks-files/tasks-files-content/tasks-files-content';
 import { TasksFilesList } from '@/components/features/tasks/tasks-files/tasks-files-list/tasks-files-list';
 import { useTasksFilesDetail } from '@/components/features/tasks/tasks-files/use-tasks-files-detail';
-import { useProjectsPageContext } from '@/components/pages/projects/providers/context';
+import { useProjectsPageContext } from '@/components/pages/projects/contexts/context';
 import { Flex } from '@/components/ui/flex';
 import { useProjectsFilesPageQuery } from '@/hooks/queries/app';
 import { getProjectsDetailId, isProjectsDetailURL, useRouter } from '@/router';
