+++ apps/nextjs/src/components/pages/workspaces/components/overview/overview.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { TasksContainer } from '@/components/features/tasks/tasks-container/tasks-container';
 import { Flex } from '@/components/ui/flex';
-import { useWorkspacesPageContext } from '../../providers/context';
+import { useWorkspacesPageContext } from '../../contexts/context';
 import { DescriptionSection } from './description-section';
 import { MembersSection } from './members-section';
 import { OverviewLeft } from './overview-left';
